+++ src/mesh/force/kantor0/imp.h
@@ -1,7 +1,7 @@
-struct double3;
+struct float3;
 namespace force_kantor0_hst {
 // tag::interface[]
-double3 dih_a(double phi, double kb, double3 a, double3 b, double3 c, double3 d);
-double3 dih_b(double phi, double kb, double3 a, double3 b, double3 c, double3 d);
+float3 dih_a(float phi, float kb, float3 a, float3 b, float3 c, float3 d);
+float3 dih_b(float phi, float kb, float3 a, float3 b, float3 c, float3 d);
 // end::interface[]
 }
